fix tf.py
