+++ neural_compressor/adaptor/tensorflow.py
@@ -1030,6 +1030,110 @@ class TensorFlowAdaptor(Adaptor):
 
         return capability
 
+    def set_tensor(self, model, tensor_dict):
+        """Quantize the bias and weight tensors in tensor_dict."""
+        from .tf_utils.graph_util import GraphAnalyzer
+
+        g = GraphAnalyzer()
+        g.graph = model.graph_def
+        graph_info = g.parse_graph()
+
+        def _get_fp32_op_name(model, tensor_name):
+            is_weight = False
+            is_biasadd = False
+            last_node_name = None
+            current_node_name = None
+            for each_node in model.graph_def.node:
+                if tensor_name in each_node.input:
+                    tensor_index = list(each_node.input).index(tensor_name)
+                    if each_node.op.find("Quantized") != -1 and tensor_index == 2:
+                        is_biasadd = True
+                        last_node_name = each_node.input[0]
+                        current_node_name = each_node.name
+
+                if tensor_name + "_qint8_const" in each_node.input:
+                    pass
+
+            return is_weight, is_biasadd, current_node_name, last_node_name
+
+        from tensorflow.core.framework import attr_value_pb2
+        from tensorflow.python.framework import dtypes, tensor_util
+
+        from neural_compressor.adaptor.tf_utils.graph_util import GraphRewriterHelper as Helper
+
+        qint32_type = dtypes.qint32.as_datatype_enum
+
+        for tensor_name, tensor_content in tensor_dict.items():
+            is_weight, is_biasadd, current_node_name, last_node_name = _get_fp32_op_name(model, tensor_name)
+
+            if is_biasadd:
+                is_biasadd_dtype_is_fp32 = graph_info[current_node_name].node.attr["Tbias"] == attr_value_pb2.AttrValue(
+                    type=dtypes.float32.as_datatype_enum
+                )
+                current_node = graph_info[current_node_name].node
+                bias_add_node = graph_info[current_node.input[2]].node
+                if is_biasadd_dtype_is_fp32:
+                    bias_add_node.attr["value"].CopyFrom(
+                        attr_value_pb2.AttrValue(
+                            tensor=tensor_util.make_tensor_proto(tensor_content, dtypes.float32, tensor_content.shape)
+                        )
+                    )
+                else:
+                    last_node = graph_info[last_node_name].node
+                    min_input = graph_info[last_node.input[-2]].node.attr["value"].tensor.float_val[0]
+                    max_input = graph_info[last_node.input[-1]].node.attr["value"].tensor.float_val[0]
+                    channel_size = tensor_content.shape[0]
+                    max_filter_node = graph_info[current_node.input[6]].node
+                    min_filter_node = graph_info[current_node.input[5]].node
+                    if max_filter_node.attr["value"].tensor.float_val:
+                        max_filter_tensor = []
+                        min_filter_tensor = []
+                        max_filter_tensor.append((max_filter_node.attr["value"].tensor.float_val)[0])
+                        min_filter_tensor.append((min_filter_node.attr["value"].tensor.float_val)[0])
+                    else:
+                        max_filter_tensor = tensor_util.MakeNdarray(min_filter_node.attr["value"].tensor)
+                        min_filter_tensor = tensor_util.MakeNdarray(min_filter_node.attr["value"].tensor)
+                    activation_range = 127.0 if current_node.attr["Tinput"].type == dtypes.qint8 else 255.0
+                    updated_bias = Helper.generate_int32_bias_for_conv(
+                        tensor_content,
+                        channel_size,
+                        max_input,
+                        min_input,
+                        max_filter_tensor,
+                        min_filter_tensor,
+                        activation_range,
+                    )
+
+                    bias_add_node.attr["dtype"].CopyFrom(attr_value_pb2.AttrValue(type=qint32_type))
+                    bias_add_node.attr["value"].CopyFrom(
+                        attr_value_pb2.AttrValue(
+                            tensor=tensor_util.make_tensor_proto(updated_bias, dtypes.int32, tensor_content.shape)
+                        )
+                    )
+                    bias_add_node.attr["value"].tensor.dtype = qint32_type
+                    current_node.attr["Tbias"].CopyFrom(attr_value_pb2.AttrValue(type=qint32_type))
+
+            if is_weight:
+                tmp_const_node = Helper.create_constant_node(
+                    current_node.name + "_weights_tmp", tensor_content.transpose(2, 3, 1, 0), dtypes.float32
+                )
+                min_filter_node = graph_info[current_node.input[5]].node
+                per_channel = True if min_filter_node.attr["value"].tensor.tensor_shape else False
+                from .tf_utils.quantize_graph_common import QuantizeGraphHelper
+
+                original_fp32_op = current_node.op.split("With")[0].split("Quantized")[-1]
+                if original_fp32_op.find("Depthwise") != -1:
+                    original_fp32_op = "DepthwiseConv2dNative"
+                qint8_const_node, min_node, max_node = QuantizeGraphHelper.generate_quantized_weight_node(
+                    original_fp32_op, tmp_const_node, per_channel
+                )
+                g.add_node(qint8_const_node, [], [current_node.name])
+                g.add_node(min_node, [], [current_node.name])
+                g.add_node(max_node, [], [current_node.name])
+                g.replace_constant_graph_with_constant_node(qint8_const_node, tensor_name)
+                g.replace_constant_graph_with_constant_node(min_node, current_node.input[5])
+                g.replace_constant_graph_with_constant_node(max_node, current_node.input[6])
+
     def inspect_weight_and_bias(self, node_list, graph_def, graph_info, graph_node_name_mapping):
         """Inspect the weights and biases."""
         import tensorflow as tf
